fix/socketIO : 최종 연결 완료
